simplify run_pytest and reduce dependencies
