+++ hi-ml-azure/src/health_azure/logging.py
@@ -1,0 +1,95 @@
+#  -------------------------------------------------------------------------------------------
+#  Copyright (c) Microsoft Corporation. All rights reserved.
+#  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+#  -------------------------------------------------------------------------------------------
+import logging
+import sys
+import time
+from contextlib import contextmanager
+from typing import Generator, Optional, Union
+
+from health_azure.utils import check_is_any_of
+
+logging_stdout_handler: Optional[logging.StreamHandler] = None
+logging_to_file_handler: Optional[logging.StreamHandler] = None
+
+
+def logging_to_stdout(log_level: Union[int, str] = logging.INFO) -> None:
+    """
+    Instructs the Python logging libraries to start writing logs to stdout up to the given logging level.
+    Logging will use a timestamp as the prefix, using UTC.
+
+    :param log_level: The logging level. All logging message with a level at or above this level will be written to
+    stdout. log_level can be numeric, or one of the pre-defined logging strings (INFO, DEBUG, ...).
+    """
+    log_level = standardize_log_level(log_level)
+    logger = logging.getLogger()
+    # This function can be called multiple times, in particular in AzureML when we first run a training job and
+    # then a couple of tests, which also often enable logging. This would then add multiple handlers, and repeated
+    # logging lines.
+    global logging_stdout_handler
+    if not logging_stdout_handler:
+        print("Setting up logging to stdout.")
+        # At startup, logging has one handler set, that writes to stderr, with a log level of 0 (logging.NOTSET)
+        if len(logger.handlers) == 1:
+            logger.removeHandler(logger.handlers[0])
+        logging_stdout_handler = logging.StreamHandler(stream=sys.stdout)
+        _add_formatter(logging_stdout_handler)
+        logger.addHandler(logging_stdout_handler)
+    print(f"Setting logging level to {log_level}")
+    logging_stdout_handler.setLevel(log_level)
+    logger.setLevel(log_level)
+
+
+def standardize_log_level(log_level: Union[int, str]) -> int:
+    """
+
+    :param log_level: integer or string (any casing) version of a log level, e.g. 20 or "INFO".
+    :return: integer version of the level; throws if the string does not name a level.
+    """
+    if isinstance(log_level, str):
+        log_level = log_level.upper()
+        check_is_any_of("log_level", log_level, logging._nameToLevel.keys())
+        return logging._nameToLevel[log_level]
+    return log_level
+
+
+def _add_formatter(handler: logging.StreamHandler) -> None:
+    """
+    Adds a logging formatter that includes the timestamp and the logging level.
+    """
+    formatter = logging.Formatter(fmt="%(asctime)s %(levelname)-8s %(message)s", datefmt="%Y-%m-%dT%H:%M:%SZ")
+    # noinspection PyTypeHints
+    formatter.converter = time.gmtime  # type: ignore
+    handler.setFormatter(formatter)
+
+
+@contextmanager
+def logging_section(gerund: str) -> Generator:
+    """
+    Context manager to print "**** STARTING: ..." and "**** FINISHED: ..." lines around sections of the log,
+    to help people locate particular sections. Usage:
+    with logging_section("doing this and that"):
+       do_this_and_that()
+
+    :param gerund: string expressing what happens in this section of the log.
+    """
+    from time import time
+
+    logging.info("")
+    msg = f"**** STARTING: {gerund} "
+    logging.info(msg + (100 - len(msg)) * "*")
+    logging.info("")
+    start_time = time()
+    yield
+    elapsed = time() - start_time
+    logging.info("")
+    if elapsed >= 3600:
+        time_expr = f"{elapsed / 3600:0.2f} hours"
+    elif elapsed >= 60:
+        time_expr = f"{elapsed / 60:0.2f} minutes"
+    else:
+        time_expr = f"{elapsed:0.2f} seconds"
+    msg = f"**** FINISHED: {gerund} after {time_expr} "
+    logging.info(msg + (100 - len(msg)) * "*")
+    logging.info("")
