+++ multimodal/environment.yml
@@ -6,4 +6,13 @@ dependencies:
   - pip=20.1.1
   - python=3.7.3
   - pip:
+    # Run requirements for hi-ml-azure
+    - azureml-sdk==1.36.0
+    - azureml-tensorboard==1.36.0
+    - conda-merge==0.1.5
+    - param==1.12
+    - pysocks==1.6.0
+    - ruamel.yaml==0.16.12
+    - tensorboard==2.6.0
+    # Test requirements
     - pytest==6.2.2
